Add files via upload
WPP_17.Write a JavaScript program to replace each character in a given string with the next in the English alphabet.  
WPP_18.Write a JavaScript program to find the maximum difference between any two adjacent elements of a given array of integers.
WPP_19.Write a JavaScript program to find the number appearing most frequently in a given array of integers.
WPP_20.Write a JavaScript program to reverse the order of bits in an integer. 
WPP_21.Write a JavaScript program to remove all characters from a given string that appear more than once.